fix imaage sizing and run all through tinypng

+++ products/views.py
@@ -70,6 +70,10 @@ def product_detail(request, product_id):
 @login_required
 def add_product(request):
     """ add a product to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the permissions required.')
+        return redirect(reverse('home'))
+    
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -91,6 +95,10 @@ def add_product(request):
 @login_required
 def edit_product(request, product_id):
     """ edit and update products"""
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the permissions required.')
+        return redirect(reverse('home'))
+    
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -115,6 +123,10 @@ def edit_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ delete a product """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the permissions required.')
+        return redirect(reverse('home'))
+    
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted')
